requirements.txt
PyYAML==6.0
gitpython==3.1.31
pinecone-client==3.0.0
requests==2.28.2
python-dotenv==1.0.0
protobuf>=4.21.6,<5.0.0
grpcio==1.54.0
grpcio-tools==1.54.0
firecrawl==2.5.4
